p4i integration Integrate 180108 from 4.0 to main: User:     paul Fixes:    Compile after change 179991
[git-p4: depot-paths = "//depot/qt/main/": change = 180110]

+++ src/gui/painting/qpaintengine_raster.cpp
@@ -2066,7 +2066,7 @@ void QRasterBuffer::prepare(QPixmap *pixmap)
 {
     prepareClip(pixmap->width(), pixmap->height());
 
-    m_buffer = (uint *)pixmap->qwsScanLine(0);
+    m_buffer = (uchar *)pixmap->qwsScanLine(0);
 
     m_width = pixmap->width();
     m_height = pixmap->height();
